requirements/base.txt
asgiref==3.11.1
attrs==25.4.0
Django==5.0
django-extensions==3.2.3
django-ratelimit==4.1.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.1.1
pydotplus==2.0.2
PyJWT==2.11.0
pyparsing==3.3.2
python-decouple==3.8
PyYAML==6.0.3
redis==7.1.1
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
